# txdxai/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: TxDxAI API
  description: Multi-tenant Cybersecurity and Ticket Automation API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        created_by_user_id:
          type: integer
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [PENDING, EXECUTED, FAILED, DERIVED]
        executed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /tickets:
    get:
      tags:
        - Tickets
      summary: Get all tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, EXECUTED, FAILED, DERIVED]
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
    
    post:
      tags:
        - Tickets
      summary: Create a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
              properties:
                subject:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  ticket:
                    $ref: '#/components/schemas/Ticket'
  /tickets/{ticket_id}:
    get:
      tags:
        - Tickets
      summary: Get ticket by ID
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
    
    put:
      tags:
        - Tickets
      summary: Update ticket
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [PENDING, EXECUTED, FAILED, DERIVED]
      responses:
        '200':
          description: Ticket updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  ticket:
                    $ref: '#/components/schemas/Ticket'
    
    delete:
      tags:
        - Tickets
      summary: Delete ticket
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ticket deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...